Update README.md
Here is a brief overview of how it works:
It sets the color of the console window to light blue on black (color 17).
It sets the console window title to "MYP Language Acquisition conceptual understanding generator by Inquiring Practitioner (2023)".
It defines a label called "loop1" and clears the console screen.
It sets a variable called "folder" to the batch file folder path.
It sets a variable called "lookup" to the path of a file called "lookup.txt" in the batch file folder.
It displays a list of critical concepts and prompts the user to select one by typing a number.
It sets a variable called "kcname" to the name of the selected key concept.
It displays a list of related concepts and prompts the user to select two by typing a number.
It sets variables called "rcname1" and "rcname2" to the names of the selected related concepts.
It generates a conceptual understanding based on the selected key concept and related concepts.

Strengths of this CMD batch file include their simplicity and ease of use. CMD is a built-in shell in Windows, so installing any software is unnecessary. It is also a very lightweight language that can be used for simple tasks. In this code, CMD is used to display text, get user input, and set variables, all of which are basic operations that CMD handles well.




The file does have several limitations compared to other languages, such as Python. It is not a programming language in the traditional sense but rather a shell language used for executing commands. 

This is my first ever code-type project on any computer OS system. It is an exciting example of how to use CMD to create a simple interactive program. However, other languages like Python may be better for my future projects requiring more advanced programming features.